rename FeatureTree to FeatureIndex
and split intersection tests into a separate file
